pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "victoruno"
version = "0.1.0"
description = "Personalized Agent to Research, Develop, and Optimize using LangChain and LangGraph with local Ollama models"
authors = [
    {name = "Cole Barbes", email = "cbarbes1@example.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "langchain>=0.1.0",
    "langgraph>=0.0.40",
    "ollama>=0.1.7",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0.3",
    "aiofiles>=23.2.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "selenium>=4.15.0",
    "playwright>=1.40.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "isort>=5.12.0",
]
gui = [
    "PyQt6>=6.5.0",
    "customtkinter>=5.2.0",
]

[project.urls]
Homepage = "https://github.com/cbarbes1/VictorUno"
Repository = "https://github.com/cbarbes1/VictorUno"
Issues = "https://github.com/cbarbes1/VictorUno/issues"

[project.scripts]
victoruno = "victoruno.cli:main"
victoruno-gui = "victoruno.gui.main:main"
victoruno-web = "victoruno.web.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true